Started filling out AnnouncementView display variables. Added variables to track id's of child Views. Changed AnnouncementView to a ConstraintLayout rather than a LinearLayout.
Added a line of code in DirectoryActivity.displayDirectory() to clear old child Views.

Started work on ScheduleActivity mock server response.

Added check in ServerRequest.request() to confirm that the variableKeys parameter is not null.

Nearly completed activity_home_page.xml. Added buttons to other activites and got the container for the AnnouncementViews set up. Presently, there is a test announcementView which was used to determine the nessisary attributes to set up in the AnnouncementView constructor.

Added background_announc_view.xml drawable object which is used as a background for AnnouncementView objects.